sched: Separate WindowSelector func
